cap: fix int truncation of cap bits >= 32 in cap_req
cap_req() declared capadd/capdel as int (32 bit) but cap bits live in
client->local->caps which is long (64 bit).  When clicap_allocate_cap
hands out bits 32-63 to a cap, OR-ing cap->cap into a 32-bit int
silently truncates them to zero -- the cap was advertised in CAP LS,
ACKed on REQ, but never actually set on the client.

The most visible victim was SASL.  Recent module additions
(obby.world/channel-bots was the straw, but anything would have done
it) pushed the sasl cap's allocated bit past 31, so HasCapability(c,
"sasl") quietly returned 0 and cmd_authenticate's first guard
silently dropped every AUTHENTICATE attempt.  CAP LS showed
sasl=PLAIN,..., CAP REQ :sasl came back with an ACK, and then the
client sat forever waiting for AUTHENTICATE +.

Widening capadd/capdel to long fixes all caps in one place.

+++ src/modules/cap.c
@@ -245,7 +245,11 @@ static void cap_req(Client *client, const char *arg)
 	ClientCapability *cap;
 	int buflen, plen;
 	int i = 0;
-	int capadd = 0, capdel = 0;
+	/* MUST be long: cap bits live in a 64-bit field (client->local->caps).
+	 * Using int here truncates any bit >= 32 to zero, which silently drops
+	 * caps registered later in module load order -- famously breaking SASL
+	 * once enough other caps have been registered ahead of it. */
+	long capadd = 0, capdel = 0;
 	int finished = 0, negate;
 	int errors = 0;
